+++ api/src/main/java/pl/kolendateam/dadcard/skills/entity/ClassSkills.java
@@ -9,6 +9,6 @@ public class ClassSkills {
 
     String nameSkill;
     boolean classSkill;
-    int skillRank;
+    double skillRank;
     
 }
